// src/app/core/restaurant_service/model/openingHoursDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TimeRange } from './timeRange';

export interface OpeningHoursDTO { 
    id?: number;
    dayOfWeek?: OpeningHoursDTO.DayOfWeekEnum;
    timeRanges?: Array<TimeRange>;
    closed?: boolean;
}
export namespace OpeningHoursDTO {
    export type DayOfWeekEnum = 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
    export const DayOfWeekEnum = {
        MONDAY: 'MONDAY' as DayOfWeekEnum,
        TUESDAY: 'TUESDAY' as DayOfWeekEnum,
        WEDNESDAY: 'WEDNESDAY' as DayOfWeekEnum,
        THURSDAY: 'THURSDAY' as DayOfWeekEnum,
        FRIDAY: 'FRIDAY' as DayOfWeekEnum,
        SATURDAY: 'SATURDAY' as DayOfWeekEnum,
        SUNDAY: 'SUNDAY' as DayOfWeekEnum
    };
}